Moved Simulator class to own file
